migrate erosion code to separate files and modules, reintegrate old code from main
