+++ .entangled/tasks.json
@@ -1,3 +1,69 @@
 {
-  "task": []
+  "task": [
+    {
+      "description": null,
+      "creates": [
+        "data/caps-osc.h5"
+      ],
+      "requires": [
+        "src/CaProd.jl"
+      ],
+      "runner": "julia",
+      "stdout": null,
+      "stdin": null,
+      "script": null,
+      "path": "examples/caps-osc.jl"
+    },
+    {
+      "description": null,
+      "creates": [
+        "docs/src/fig/b13-capsosc-crosssection.png"
+      ],
+      "requires": [
+        "data/caps-osc.h5",
+        "ext/VisualizationExt.jl"
+      ],
+      "runner": "julia",
+      "stdout": null,
+      "stdin": null,
+      "script": null,
+      "path": "examples/plot-caps-osc.jl"
+    },
+    {
+      "description": null,
+      "creates": [
+        "data/ca-prod-slope.h5"
+      ],
+      "requires": [
+        "src/CaProd.jl"
+      ],
+      "runner": "julia",
+      "stdout": null,
+      "stdin": null,
+      "script": null,
+      "path": "examples/production-only/cap-slope.jl"
+    },
+    {
+      "description": null,
+      "creates": [
+        "docs/src/fig/b13-crosssection.png"
+      ],
+      "requires": [
+        "data/ca-prod-slope.h5",
+        "ext/VisualizationExt.jl"
+      ],
+      "runner": "julia",
+      "stdout": null,
+      "stdin": null,
+      "script": null,
+      "path": "examples/production-only/plot-cap-slope.jl"
+    },
+    {
+      "name": "figures",
+      "requires": [
+        "docs/src/fig/b13-capsosc-crosssection.png",
+        "docs/src/fig/b13-crosssection.png"
+      ]
+    }
+  ]
 }
