+++ examples/caps-osc.jl
@@ -1,31 +1,25 @@
 # ~/~ begin <<docs/src/ca-with-production.md#examples/caps-osc.jl>>[init]
-using CarboKitten
-using CarboKitten.CaProd
-using CarboKitten.Burgess2013
+#| creates: data/caps-osc.h5
+#| requires: src/CaProd.jl
 
-#change sinouid function
-DEFAULT_INPUT = CarboKitten.CaProd.Input(
-    sea_level = t -> 20 * sin(2π * t) + 5 * sin(2π * t / 0.112), 
+using CarboKitten.CaProd
+
+DEFAULT_INPUT = CaProd.Input(
+    sea_level = t -> 4 * sin(2π * t / 0.2), 
     subsidence_rate = 50.0,
     initial_depth = x -> x / 2,
     grid_size = (50, 100),
     phys_scale = 1.0,
-    Δt = 0.001,
-    write_interval = 1,
-    time_steps = 2000,
+    Δt = 0.0001,
+    write_interval = 10,
+    time_steps = 10000,
     facies = [
-        Burgess2013.Facies((4, 10), (6, 10), 500.0, 0.8, 300, 1000, 2730, 0.5),
-        Burgess2013.Facies((4, 10), (6, 10), 400.0, 0.1, 300, 1000, 2730, 0.5),
-        Burgess2013.Facies((4, 10), (6, 10), 100.0, 0.005, 300, 1000, 2730, 0.5)
+        CaProd.Facies((4, 10), (6, 10), 500.0, 0.8, 300),
+        CaProd.Facies((4, 10), (6, 10), 400.0, 0.1, 300),
+        CaProd.Facies((4, 10), (6, 10), 100.0, 0.005, 300)
     ],
-    insolation = 2000.0,
-    temp = 288.0,
-    precip = 1000.0,
-    pco2 = 10^(-1.5),
-    alpha = 2e-3,
-    erosion_type = 1
-
+    insolation = 2000.0
 )
 
-CarboKitten.CaProd.main(DEFAULT_INPUT, "data/caps-osc.h5")
+CaProd.main(DEFAULT_INPUT, "data/caps-osc.h5")
 # ~/~ end
