+++ examples/plot-caps-osc.jl
@@ -1,7 +1,12 @@
 # ~/~ begin <<docs/src/ca-with-production.md#examples/plot-caps-osc.jl>>[init]
+#| creates: docs/src/fig/b13-capsosc-crosssection.png
+#| requires: data/caps-osc.h5 ext/VisualizationExt.jl
+#| collect: figures
+
 module Script
+    using CairoMakie
+    using GeometryBasics
     using CarboKitten.Visualization
-    using GLMakie
 
     function main()
         f = Figure()
