+++ examples/production-only/cap-slope.jl
@@ -1,4 +1,7 @@
-# ~/~ begin <<docs/src/ca-with-production.md#examples/cap-slope.jl>>[init]
+# ~/~ begin <<docs/src/ca-with-production.md#examples/production-only/cap-slope.jl>>[init]
+#| creates: data/ca-prod-slope.h5
+#| requires: src/CaProd.jl
+
 using CarboKitten.CaProd
 
 DEFAULT_INPUT = CaProd.Input(
@@ -11,9 +14,9 @@ DEFAULT_INPUT = CaProd.Input(
     write_interval=1,
     time_steps=1000,
     facies=[
-        CaProd.Facies((4, 10), (6, 10), 500.0, 0.8, 300, 1000, 2730, 0.5),
-        CaProd.Facies((4, 10), (6, 10), 400.0, 0.1, 300, 1000, 2730, 0.5),
-        CaProd.Facies((4, 10), (6, 10), 100.0, 0.005, 300, 1000, 2730, 0.5)
+        CaProd.Facies((4, 10), (6, 10), 500.0, 0.8, 300),
+        CaProd.Facies((4, 10), (6, 10), 400.0, 0.1, 300),
+        CaProd.Facies((4, 10), (6, 10), 100.0, 0.005, 300)
     ],
     insolation=2000.0
 )
